🏗️ Update FSD architecture
