+++ src/__tests__/advanced.test.tsx
@@ -1,8 +1,8 @@
 import { fireEvent, render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import App from "../App";
-import * as utils from "../utils";
+import App from "../apps/App";
+import * as utils from "../shared/utils";
 
 const renderLogMock = vi.spyOn(utils, "renderLog");
 const generateItemsSpy = vi.spyOn(utils, "generateItems");
