+++ src/apps/layout/CommonLayout.tsx
@@ -1,0 +1,12 @@
+import { useTheme } from "../../shared/theme";
+
+export const CommonLayout = ({ children }: { children: React.ReactNode }) => {
+  const { theme } = useTheme();
+  return (
+    <div
+      className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
+    >
+      {children}
+    </div>
+  );
+};
