+++ src/apps/layout/index.ts
@@ -1,0 +1,1 @@
+export { CommonLayout } from "./CommonLayout";
