+++ src/apps/providers/authentication/authentication.provider.tsx
@@ -1,0 +1,39 @@
+import { ReactNode, useState } from "react";
+import { useCallback, useMemo } from "../../../@lib";
+import { AuthenticationContext } from "../../../entities/authentication";
+import { User } from "../../../entities/user";
+import { useNotification } from "../../../features/notification";
+
+export const AuthenticationProvider: React.FC<{ children: ReactNode }> = ({
+  children,
+}) => {
+  const [user, setUser] = useState<User | null>(null);
+  const { addNotification } = useNotification();
+  const login = useCallback(
+    (email: string) => {
+      setUser({ id: 1, name: "홍길동", email });
+      addNotification("성공적으로 로그인되었습니다", "success");
+    },
+    [addNotification],
+  );
+
+  const logout = useCallback(() => {
+    setUser(null);
+    addNotification("로그아웃되었습니다", "info");
+  }, [addNotification]);
+
+  const authenticationContextValue = useMemo(
+    () => ({
+      user,
+      login,
+      logout,
+    }),
+    [user, login, logout],
+  );
+
+  return (
+    <AuthenticationContext.Provider value={authenticationContextValue}>
+      {children}
+    </AuthenticationContext.Provider>
+  );
+};
