+++ src/apps/providers/authentication/index.ts
@@ -1,0 +1,1 @@
+export { AuthenticationProvider } from "./authentication.provider";
