+++ src/apps/providers/notification/index.ts
@@ -1,0 +1,1 @@
+export { NotificationProvider } from "./notification.provider";
