+++ src/apps/providers/notification/notification.provider.tsx
@@ -1,0 +1,40 @@
+import { ReactNode, useState } from "react";
+import { useCallback, useMemo } from "../../../@lib";
+import {
+  Notification,
+  NotificationContext,
+} from "../../../entities/notification";
+
+export const NotificationProvider: React.FC<{ children: ReactNode }> = ({
+  children,
+}) => {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const addNotification = useCallback(
+    (message: string, type: Notification["type"]) => {
+      setNotifications((prev) => [...prev, { id: Date.now(), message, type }]);
+    },
+    [],
+  );
+
+  const removeNotification = useCallback((id: number) => {
+    setNotifications((prev) =>
+      prev.filter((notification) => notification.id !== id),
+    );
+  }, []);
+
+  const notificationContextValue = useMemo(
+    () => ({
+      notifications,
+      addNotification,
+      removeNotification,
+    }),
+    [notifications, addNotification, removeNotification],
+  );
+
+  return (
+    <NotificationContext.Provider value={notificationContextValue}>
+      {children}
+    </NotificationContext.Provider>
+  );
+};
