+++ src/apps/providers/theme/index.ts
@@ -1,0 +1,1 @@
+export { ThemeProvider } from "./theme.provider";
