+++ src/apps/providers/theme/theme.provider.tsx
@@ -1,0 +1,18 @@
+import { ReactNode, useState } from "react";
+import { ThemeContext } from "../../../shared/theme";
+
+export const ThemeProvider: React.FC<{ children: ReactNode }> = ({
+  children,
+}) => {
+  const [theme, setTheme] = useState("light");
+
+  const toggleTheme = () => {
+    setTheme((prev) => (prev === "light" ? "dark" : "light"));
+  };
+
+  return (
+    <ThemeContext.Provider value={{ theme, toggleTheme }}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
