+++ src/entities/authentication/index.ts
@@ -1,0 +1,4 @@
+export {
+  AuthenticationContext,
+  type AuthenticationContextType,
+} from "./model/authentication.types";
