+++ src/entities/notification/index.ts
@@ -1,0 +1,5 @@
+export {
+  NotificationContext,
+  type Notification,
+  type NotificationContextType,
+} from "./model/notification.types";
