+++ src/entities/product/index.ts
@@ -1,0 +1,2 @@
+export { type Product } from "./model/product.types";
+export { Item } from "./ui/Item";
