+++ src/entities/product/model/product.types.ts
@@ -1,0 +1,6 @@
+export interface Product {
+  id: number;
+  name: string;
+  category: string;
+  price: number;
+}
