+++ src/features/authentication/hooks/useAuthentication.ts
@@ -1,0 +1,12 @@
+import { useContext } from "react";
+import { AuthenticationContext } from "../../../entities/authentication";
+
+export const useAuthentication = () => {
+  const context = useContext(AuthenticationContext);
+  if (context === undefined) {
+    throw new Error(
+      "useAuthentication must be used within an AuthenticationProvider",
+    );
+  }
+  return context;
+};
