+++ src/features/authentication/index.ts
@@ -1,0 +1,2 @@
+export { useAuthentication } from "./hooks/useAuthentication";
+export { AuthenticationService } from "./service/authentication.service";
