+++ src/features/authentication/service/authentication.service.ts
@@ -1,0 +1,11 @@
+export const AuthenticationService = () => ({
+  login: (
+    email: string,
+    //  password: string
+  ) => {
+    return { id: 1, name: "홍길동", email };
+  },
+  logout: () => {
+    return null;
+  },
+});
