+++ src/features/notification/hooks/useNotification.ts
@@ -1,0 +1,12 @@
+import React from "react";
+import { NotificationContext } from "../../../entities/notification";
+
+export const useNotification = () => {
+  const context = React.useContext(NotificationContext);
+  if (context === undefined) {
+    throw new Error(
+      "useNotification must be used within a NotificationProvider",
+    );
+  }
+  return context;
+};
