+++ src/shared/theme/hooks/useTheme.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { ThemeContext } from "../model/theme.model";
+
+export const useTheme = () => {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error("useTheme must be used within a ThemeProvider");
+  }
+  return context;
+};
