+++ src/shared/theme/index.ts
@@ -1,0 +1,2 @@
+export { useTheme } from "./hooks/useTheme";
+export { ThemeContext, type ThemeContextType } from "./model/theme.model";
