+++ src/widgets/form/index.ts
@@ -1,0 +1,1 @@
+export { ComplexForm } from "./ComplexForm";
