+++ src/widgets/notification/index.ts
@@ -1,0 +1,1 @@
+export { NotificationSystem } from "./NotificationSystem";
