+++ src/widgets/products/index.ts
@@ -1,0 +1,1 @@
+export { ProductList } from "./ProductList";
